requirements.txt
Flask>=2.3.0
Gunicorn>=21.2.0
python-dotenv>=0.19.0
Flask-Cors>=4.0.0
Flask-Limiter>=4.0.0
markupsafe>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

